============ Schedulers ==============
--------------Termination----------------
There is this mess where scheduler and engine has to cooperate with regards to termination evaluation.
And the whole debate about NEWTASK / WAITING / COMPLETE.

I factored out the the termination evalation so that:

 -- both the shared_terminator and the task_count_terminator are compatible. (common interface, and behave the same way)
    As a result, the "Waiting" mechanism is managed by the engine and not the scheduler.
    However, since each scheduler has it own "preferred" terminator:

REQUIREMENT:
 -- Each scheduler exposes
    1: a typedef for a "terminator_type"
    2: a function terminator_type& get_terminator();

-- get_task() now only returns NEWTASK or EMPTY.

It is now up to the engine, when seeing EMPTY, perform termination evaluation using the terminator returned by scheduler->get_terminator()

ADDITIONS:
 -- For Schedulers which do not require the complex shared/task_count terminators,
    I added a new terminator called controlled_termination
    It basically always returns false until a flag is set.



----------Scheduler Construction-----------
Scheduler construction is now delayed to "as late as possible".
Essentially, it is only constructed on engine::start() or on first call to engine::get_scheduler()
and is deleted when engine::start() completes.


----------Scheduler Options-----------
REQUIREMENT:
To aid in scheduler specific command line parsing,
each scheduler must expose 2 functions
    void parse_options(std::stringstream &strm);
    void print_options_help();

For instance:
    void sweep_scheduler::print_options_help() {
      std::cout << "sweep(ordering)\n";
      std::cout << "ordering: linear, permute\n";
    }

    void parse_options(std::stringstream &strm) {
      std::string ordertype; strm >> ordertype;
       .....
    }

ADDITIONS:
If possible, setting scheduler options should not require construction of the graph.
Current situation is a little complicated since the engine_factory could call
engine->get_scheduler() and set options in it, even before the graph is really ready.
There are other possible breaking interactions of this type when working with the core.

Instead, I defined a class in schedulers/support called scheduler_option_cache
The scheduler_option_cache essentially manages

  - std::stringstream cmdoptions;
    These are unparsed options passed on the command line
  -  std::vector<std::pair<graphlab::scheduler_options::options_enum, void*> > options;
     These are options that can be passed via scheduler::set_option

Both core and engine will have an instance of an scheduler_option_cache
Parameters passed on the command line will be automatically inserted into the cache.
To set scheduler options programmatically, the user calls:
 - core.set_sched_options(option, value)
 or
 - engine.set_sched_options(option,value)

These scheduler options are merged into the engine, and are applied when constructing the scheduler and when calling engine.start()


-------- Various Bug Fixes -----------
 -- Colored scheduler could lose up to #ncpus tasks
 -- round_robin scheduler #iterations is incorrect when start vertex is not 0.
 -- round robin scheduler has a race condition where actual task counts could be slightly above desired counts
 -- sweep scheduler now has "permute" option

-------------REMOVALS -------------------
 -- "multiqueue" scheduler (same as multiqueue fifo
 -- "threaded" engine (we have async)
 -- "async_sim" engine (async does async_sim)
 -- set_scheduler
 -- splash scheduler